pc - temporarily comment out jacoco in pom.xml
There is currently an unknown bug that is causing mvn test to fail
with jacoco enabled in the pom.xml.  Until we can determine what is
causing that, we need to remove jacoco.    That will enable students
to make progress with the project while still having test coverage
on GitHub actions.